Moving generic scheme interaction functions to new file
